+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+New features:
+ * `Object.ExpirationEpoch`/`Object.SetExpirationEpoch` methods for version-independent access to the expiration epoch system attribute (#254)
+
 ## [1.0.0-rc.22] - 2026-09-01
 
 API 2.26 compatibility and optimizations, upgrade your applications to make